+++ src/components/pages/login/LoginPage.tsx
@@ -2,8 +2,8 @@ import styled from "styled-components";
 import TitleLogin from "./formElement/TitleLogin";
 import { useState } from "react";
 import SwitchLogin from "./formElement/SwitchLogin";
-import LoginForm from "./LoginForm";
-import RegisterForm from "./registerForm";
+import LoginForm from "./loginElement/LoginForm";
+import RegisterForm from "./loginElement/registerForm";
 
 export default function LoginPage() {
   const [isRegister, setIsRegister] = useState(false);
